Consolidate model namespaces into models and runtime
Move MoE execution backends from src/moe into src/runtime/moe.
Move MoE/model spec framework from src/moe_model into src/models/moe.
Move DeepSeek and MiniMax architecture specs under src/models.
Split MiniMax-M2 MoE planning from runtime loader/cache code.
Move DeepSeek loader and partition helpers under src/runtime/deepseek_v4.
Remove legacy src/moe, src/moe_model, and runtime transformer/loader/policy shims.

Validation:
- 69 passed, 1 skipped focused/broader pytest regression
- MiniMax real GGUF inspect and routed block check OK
- DeepSeek real GGUF inspect validation OK
- fused attention script smoke OK
