+++ setup.py
@@ -27,8 +27,13 @@ setup(
         "src.encoding",
         "src.gguf",
         "src.kernels",
-        "src.moe",
+        "src.models",
+        "src.models.deepseek_v4",
+        "src.models.minimax_m2",
+        "src.models.moe",
         "src.runtime",
+        "src.runtime.deepseek_v4",
+        "src.runtime.moe",
         "src.server",
     ],
     ext_modules=[
