+++ src/models/minimax_m2/__init__.py
@@ -1,0 +1,3 @@
+"""MiniMax-M2 model-shape implementation for PocketMoE."""
+
+ARCHITECTURE = "minimax-m2"
