+++ src/models/moe/__init__.py
@@ -1,8 +1,11 @@
-from src.moe_model.registry import detect_spec, get_spec, known_architectures
-from src.moe_model.spec import (
+"""Common MoE model spec framework."""
+
+from src.models.moe.registry import detect_spec, get_spec, known_architectures
+from src.models.moe.spec import (
     CapabilityItem,
     CapabilityReport,
     MoEArchitectureParams,
+    MoEModelSpec,
     PlacementDecision,
     SpecValidation,
     TensorMapping,
@@ -12,6 +15,7 @@ __all__ = [
     "CapabilityItem",
     "CapabilityReport",
     "MoEArchitectureParams",
+    "MoEModelSpec",
     "PlacementDecision",
     "SpecValidation",
     "TensorMapping",
