+++ src/runtime/deepseek_v4/__init__.py
@@ -1,0 +1,1 @@
+"""DeepSeek-V4 runtime utilities."""
