+++ src/runtime/moe/__init__.py
@@ -1,0 +1,1 @@
+"""Common MoE execution backends for PocketMoE runtime."""
