+++ tests/test_minimax_m2_spec.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from src.gguf.bundle import read_gguf_bundle
-from src.moe_model.minimax_m2_spec import MiniMaxM2Spec
+from src.models.minimax_m2.spec import MiniMaxM2Spec
 from tests.gguf_test_utils import GGML_F32, write_gguf, write_minimax_bundle
