+++ tests/test_moe_model_registry.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from src.gguf.bundle import read_gguf_bundle
-from src.moe_model.registry import detect_spec, get_spec, known_architectures
+from src.models.moe.registry import detect_spec, get_spec, known_architectures
 from tests.gguf_test_utils import write_gguf, write_minimax_bundle
